Creating the Capture Table

+++ src/main/java/com/endava/pokemonChallengue/models/dto/login/LogInDto.java
@@ -10,9 +10,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Builder
 public class LogInDto {
-    private Long id;
+    private int id;
     private String email;
     private String username;
-
-
 }
